Refactor: ORM
schema, typeDefs and query rewrite
