+++ api/schemas/schema.js
@@ -1,68 +1,84 @@
-const { gql } = require('apollo-server-express');
-
-
-const typeDefs = gql`
-  type Query {
-    launches(
-      """
-      The number of results to show. Must be >= 1. Default = 20
-      """
-      pageSize: Int
-      """
-      If you add a cursor here, it will only return results _after_ this cursor
-      """
-      after: String
-    ): LaunchConnection!
-    launch(id: ID!): Launch
-    me: User
-  }
-  type Mutation {
-    # if false, signup failed -- check errors
-    bookTrips(launchIds: [ID]!): TripUpdateResponse!
-    # if false, cancellation failed -- check errors
-    cancelTrip(launchId: ID!): TripUpdateResponse!
-    login(email: String): String # login token
-  }
-  type TripUpdateResponse {
-    success: Boolean!
-    message: String
-    launches: [Launch]
-  }
-  """
-  Simple wrapper around our list of launches that contains a cursor to the
-  last item in the list. Pass this cursor to the launches query to fetch results
-  after these.
-  """
-  type LaunchConnection {
-    cursor: String!
-    hasMore: Boolean!
-    launches: [Launch]!
-  }
-  type Launch {
-    id: ID!
-    site: String
-    mission: Mission
-    rocket: Rocket
-    isBooked: Boolean!
-  }
-  type Rocket {
-    id: ID!
-    name: String
+module.exports = `
+  input UserInput {
+    username: String
+    email: String
+    state: String
     type: String
   }
-  type User {
-    id: ID!
-    email: String!
-    trips: [Launch]!
+
+  input RoleInput {
+    app_id: Int!
+    name: String!
+    description: String!
+    state: String
   }
-  type Mission {
+
+  input UserRoleInput {
+    role_id: Int!
+    user_id: Int!
+    state: String
+  }
+
+  input AppInput {
+    name: String!
+    description: String!
+  }
+
+  type UserRole {
+    role_id: Int
+    user_id: Int
+    state: String
+  }
+
+
+  type Role {
+    id: Int
+    app_id: Int
     name: String
-    missionPatch(size: PatchSize): String
+    description: String
+    state: String
   }
-  enum PatchSize {
-    SMALL
-    LARGE
+
+  type User {
+    id: Int
+    user_id: Int
+    username: String
+    email: String
+    state: String
+    type: String
+    roles:[Role]
+  }
+
+  type App {
+    id: Int
+    app_id: Int
+    name: String
+    description: String
+  }
+
+  type Query {
+    getUser(user_id: ID!): User
+    getUserRole(user_id:ID!): [UserRole]
+    getApp(app_id: ID): App
+    getRole(app_id: ID, role_id: ID): [Role]
+    users:[User]
+    roles:[Role]
+    apps:[App]
+  }
+
+  type Mutation {
+    createUser(input: UserInput): User
+    createApp(input: AppInput): App
+    createRole(input: RoleInput): Role
+    createUserRole(input: UserRoleInput): UserRole
+    updateUser(user_id: ID!, input: UserInput): User
+    updateApp(app_id: ID!): App
+    updateRole(role_id: ID!): Role
+    
+  }
+
+  schema {
+    query: Query
+    mutation: Mutation
   }
 `;
-
-module.exports = typeDefs;
